# Conflicts:
#	src/main/java/com/example/nyabankomaten/MainController.java

+++ src/main/java/com/example/nyabankomaten/Account.java
@@ -49,10 +49,10 @@ public class Account implements Printable {
         balance -= loan;
         loan = 0;
     }
-    public void deposit(int deposit){
+    public void deposit(long deposit){
         balance += deposit;
     }
-    public void withdraw(int withdraw){
+    public void withdraw(long withdraw){
         if(withdraw <= balance){
             balance -= withdraw;
         }
